+++ v0/js/tool.js
@@ -121,11 +121,13 @@ export class Tool {
         backgroundColor: tool.backgroundColor,
         ligandColor: tool.ligandColor,
         ligandElement: tool.ligandElement,
+        proteinElement: tool.proteinElement,
         backgroundOpacity: tool.backgroundOpacity,
         proteinOpacity: tool.proteinOpacity,
         selectionOpacity: tool.selectionOpacity,
         showGlycans: tool.showGlycans,
         showNucleotides: tool.showNucleotides,
+        showNonCarbonHydrogens: tool.showNonCarbonHydrogens,
       },
       tool.data
     );
@@ -167,10 +169,15 @@ export class Tool {
     );
     tool.initCheckbox(d3.select("#showGlycans"), tool.showGlycans);
     tool.initCheckbox(d3.select("#showNucleotides"), tool.showNucleotides);
+    tool.initCheckbox(
+      d3.select("#showNonCarbonHydrogens"),
+      tool.showNonCarbonHydrogens
+    );
     tool.initColorPicker(d3.select("#proteinColor"), tool.proteinColor);
     tool.initColorPicker(d3.select("#backgroundColor"), tool.backgroundColor);
     tool.initColorPicker(d3.select("#ligandColor"), tool.ligandColor);
     tool.initCheckbox(d3.select("#ligandElement"), tool.ligandElement);
+    tool.initCheckbox(d3.select("#proteinElement"), tool.proteinElement);
     tool.initRange(d3.select("#proteinOpacity"), tool.proteinOpacity);
     tool.initRange(d3.select("#selectionOpacity"), tool.selectionOpacity);
     tool.initRange(d3.select("#backgroundOpacity"), tool.backgroundOpacity);
@@ -532,11 +539,13 @@ export class Tool {
       backgroundColor: { abbrev: "bc", default: "#d3d3d3", json: false },
       ligandColor: { abbrev: "lc", default: "#d3d3d3", json: false },
       ligandElement: { abbrev: "le", default: false, json: false },
+      proteinElement: { abbrev: "pce", default: false, json: false },
       proteinOpacity: { abbrev: "po", default: "1", json: false },
       selectionOpacity: { abbrev: "so", default: "1", json: false },
       backgroundOpacity: { abbrev: "bo", default: "1", json: false },
       showGlycans: { abbrev: "g", default: false, json: false },
       showNucleotides: { abbrev: "n", default: false, json: false },
+      showNonCarbonHydrogens: { abbrev: "h", default: false, json: false },
       filters: {
         abbrev: "fi",
         default: tool.data[tool.dataset].filter_cols